Bump MC version to 26.2
