--- deobsf/v26_1/src/main/resources/pcf.mixins.v26_1.json
@@ -1,18 +1,0 @@
-{
-  "compatibilityLevel": "JAVA_8",
-  "minVersion": "0.8",
-  "injectors": {
-    "defaultRequire": 1
-  },
-  "required": true,
-  "plugin": "org.adde0109.pcf.mixin.plugin.PCFMixinPlugin",
-  "package": "org.adde0109.pcf.mixin",
-  "server": [
-    "v26_1.crossstich.ArgumentNodeStubMixin",
-    "v26_1.crossstich.ArgumentTypeInfoMixin",
-    "v26_1.forwarding.ConnectionMixin",
-    "v26_1.forwarding.ServerLoginPacketListenerImplMixin",
-    "v26_1.forwarding.ServerLoginPacketListenerImplHelloHybridMixin",
-    "v26_1.forwarding.ServerLoginPacketListenerImplHelloMixin"
-  ]
-}
