+++ deobsf/v26_2/gradle.properties
@@ -1,0 +1,3 @@
+minecraft_version=26.2
+forge_version=65.1.0
+neoforge_version=37-beta
