+++ deobsf/v26_2/src/forge/java/org/adde0109/pcf/v26_2/forge/ReloadConfigInit.java
@@ -1,4 +1,4 @@
-package org.adde0109.pcf.v20_4.forge;
+package org.adde0109.pcf.v26_2.forge;
 
 import dev.neuralnexus.taterapi.Wrapped;
 import dev.neuralnexus.taterapi.meta.MetaAPI;
@@ -7,27 +7,25 @@ import dev.neuralnexus.taterapi.meta.anno.Versions;
 import dev.neuralnexus.taterapi.meta.enums.MinecraftVersion;
 import dev.neuralnexus.taterapi.meta.enums.Platform;
 
-import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.eventbus.api.bus.BusGroup;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
 import net.minecraftforge.fml.javafmlmod.FMLModContainer;
 
 import org.adde0109.pcf.PCF;
 import org.adde0109.pcf.PCFInitializer;
-import org.adde0109.pcf.v26_2.forge.Config;
 
-@AConstraint(
-        platform = Platform.FORGE,
-        version = @Versions(min = MinecraftVersion.V17, max = MinecraftVersion.V21_5))
+@AConstraint(platform = Platform.FORGE, version = @Versions(min = MinecraftVersion.V21_6))
 public final class ReloadConfigInit implements PCFInitializer {
     @Override
     public void onInit() {
-        final IEventBus eventBus =
+        final BusGroup modBusGroup =
                 MetaAPI.instance()
                         .meta()
                         .<FMLModContainer>mod(PCF.MOD_ID)
                         .map(Wrapped::unwrap)
-                        .map(FMLModContainer::getEventBus)
+                        .map(FMLModContainer::getModBusGroup)
                         .orElseThrow();
-        eventBus.addListener((ModConfigEvent.Reloading _) -> Config.reload());
+        ModConfigEvent.Reloading.getBus(modBusGroup)
+                .addListener((ModConfigEvent.Reloading _) -> Config.reload());
     }
 }
