+++ modern/v14_4/build.gradle.kts
@@ -21,7 +21,7 @@ unimined.minecraft(forge) {
 
 dependencies {
     forgeCompileOnly(libs.mixin)
-    forgeCompileOnly(srcSetAsDep(":deobsf:v26_1", "forge"))
+    forgeCompileOnly(srcSetAsDep(":deobsf:v26_2", "forge"))
     forgeCompileOnly(project(":common"))
 }
