+++ modern/v16_5/src/forge/java/org/adde0109/pcf/v16_5/forge/ReloadConfigInit.java
@@ -13,6 +13,7 @@ import net.minecraftforge.fml.javafmlmod.FMLModContainer;
 
 import org.adde0109.pcf.PCF;
 import org.adde0109.pcf.PCFInitializer;
+import org.adde0109.pcf.v26_2.forge.Config;
 
 @AConstraint(
         platform = Platform.FORGE,
@@ -27,7 +28,6 @@ public final class ReloadConfigInit implements PCFInitializer {
                         .map(Wrapped::unwrap)
                         .map(FMLModContainer::getEventBus)
                         .orElseThrow();
-        eventBus.addListener(
-                (ModConfig.Reloading _) -> org.adde0109.pcf.v26_1.forge.Config.reload());
+        eventBus.addListener((ModConfig.Reloading _) -> Config.reload());
     }
 }
