+++ modern/v17_1/build.gradle.kts
@@ -20,7 +20,7 @@ unimined.minecraft(forge) {
 }
 
 dependencies {
-    forgeCompileOnly(srcSetAsDep(":deobsf:v26_1", "forge"))
+    forgeCompileOnly(srcSetAsDep(":deobsf:v26_2", "forge"))
     forgeCompileOnly(project(":common"))
 }
