+++ test/build.gradle.kts
@@ -23,18 +23,18 @@ val versions: Map<String, List<String>> = mapOf(
         "1.14.4", "1.15.2", "1.16.5",
         "1.17.1", "1.18.2", "1.19", "1.19.2", "1.19.4", "1.20.1", "1.20.2", "1.20.4",
         "1.21.1", "1.21.5",
-        "26.1.2"
+        "26.2"
     ),
     "neoforge" to listOf(
         "1.20.2", "1.20.4", "1.21.1", "1.21.5",
-        "26.1.2"
+        "26.2"
     ),
     "vanilla" to listOf(
         "1.12.2", "1.13.2",
         "1.14.4", "1.15.2", "1.16.5",
         "1.17.1", "1.18.2", "1.19", "1.19.2", "1.19.4", "1.20.1", "1.20.2", "1.20.4",
         "1.21.1", "1.21.5",
-        "26.1.2"
+        "26.2"
     )
 )
